+++ srcs/add_int_prefix.c
@@ -6,13 +6,13 @@
 /*   By: tkhemniw <gt.khemniwat@gmail.com>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/05 13:00:35 by tkhemniw          #+#    #+#             */
-/*   Updated: 2022/09/08 20:27:22 by tkhemniw         ###   ########.fr       */
+/*   Updated: 2022/09/09 10:23:44 by tkhemniw         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/ft_printf.h"
 
-void	add_int_prefix(char **str, int d, t_format *f)
+void	add_int_prefix(char **str, long d, t_format *f)
 {
 	if (d < 0)
 		**str = '-';
